+++ src/routes/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from "react-router-dom";
-import { useUserStore } from "../../store/useUserStore";
+import { useUserStore } from "../store/useUserStore";
 
 export const ProtectedRoute = () => {
   const user = useUserStore((state) => state.user);
