+++ src/routes/index.tsx
@@ -1,6 +1,6 @@
 // src/routes/index.tsx
 import { Routes, Route } from "react-router-dom";
-import { ProtectedRoute } from "../components/auth/ProtectedRoute";
+import { ProtectedRoute } from "./ProtectedRoute";
 
 import Login from "../pages/Login";
 import Signup from "../pages/Signup";
@@ -29,12 +29,12 @@ const AppRoutes = () => {
 
         {/* 로그인 필요한 페이지들 */}
         {/* <Route element={<ProtectedRoute />}> */}
-          <Route path="/user/edit" element={<EditProfile />} />
-          <Route path="/playlist/create" element={<PlaylistCreate />} />
-          <Route path="/" element={<Home />} />
-          <Route path="/mypage" element={<MyPage />} />
-          <Route path="/subscriptions" element={<Subscriptions />} />
-          <Route path="/playlist/:id" element={<PlaylistDetail />} />
+        <Route path="/user/edit" element={<EditProfile />} />
+        <Route path="/playlist/create" element={<PlaylistCreate />} />
+        <Route path="/" element={<Home />} />
+        <Route path="/mypage" element={<MyPage />} />
+        <Route path="/subscriptions" element={<Subscriptions />} />
+        <Route path="/playlist/:id" element={<PlaylistDetail />} />
         {/* </Route> */}
       </Route>
     </Routes>
